requirements.txt
aio-pika>=8.0.0
aiohttp>=3.8.0
pydantic>=1.10.0
jsonschema>=4.0.0
prometheus-client>=0.14.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-otlp>=1.18.0
pytest>=7.0.0
pytest-aiohttp>=1.0.0